Relocate create scraper overlay component
